Set size according to viewport in component AspectAnim
The original idea was to let the outside parent determine the maximum size to adapt the aspect video, but it's not required because I have the `Wrap` component now and CSS can only select the screen size queries anyway. Letting the aspect component be always fullscreen is fine since it will be reused in other components like `Presentation` that can go inline-fullscreen or totally fullscreen while keeping the ratio.