Clean project and create tsd-web---mvp as Gradle Basic structure
The project `tsd-web---mvp` will be a multimodule Gradle project, and the branch `ops` is creating the project from clean start to keep everything in sync compared to the migrating structure of the TSD Web Prototype. That requires to move existing project files to clean the directory and run `gradle init` with the `4: Basic (build structure only)` option to commit the root project `tsd-web---mvp`, which is neither an app nor library itself but the multimodule project root.